Move Comments to usecase package
