+++ WordPress/src/main/java/org/wordpress/android/ui/stats/refresh/usecases/InsightsUseCase.kt
@@ -10,7 +10,6 @@ import org.wordpress.android.fluxc.store.StatsStore
 import org.wordpress.android.fluxc.store.StatsStore.InsightsTypes
 import org.wordpress.android.modules.BG_THREAD
 import org.wordpress.android.modules.UI_THREAD
-import org.wordpress.android.ui.stats.refresh.CommentsUseCase
 import org.wordpress.android.ui.stats.refresh.InsightsItem
 import org.wordpress.android.ui.stats.refresh.NavigationTarget
 import org.wordpress.android.util.combineMap
